update env variables for
