+++ apps/admin/.env.example
@@ -7,13 +7,11 @@ DATABASE_URL=postgresql://USER:PASSWORD@localhost:5432/zoonk
 # Direct database connection URL for Prisma Migrate
 DATABASE_URL_UNPOOLED=
 
-# This app's URL (used for building callback URLs)
-# Use http://localhost:3001 for local development
-NEXT_PUBLIC_ADMIN_APP_URL=https://admin.zoonk.com
+# This app's domain (eg localhost:3001, admin.zoonk.com)
+NEXT_PUBLIC_APP_DOMAIN=localhost:3001
 
 # Centralized Auth App URL
-# Use http://localhost:3004 for local development
-NEXT_PUBLIC_AUTH_APP_URL=https://auth.zoonk.com
+NEXT_PUBLIC_AUTH_APP_URL=http://localhost:3004
 
 ### OPTIONAL VARIABLES ###
