+++ apps/admin/src/app/(auth)/login/page.tsx
@@ -1,12 +1,9 @@
 import { getSession } from "@zoonk/core/users";
 import { FullPageLoading } from "@zoonk/ui/components/loading";
-import { buildAuthLoginUrl } from "@zoonk/utils/auth-url";
+import { buildAuthLoginUrl, getBaseUrl } from "@zoonk/utils/url";
 import { redirect } from "next/navigation";
 import { Suspense } from "react";
 import { LoginRedirect } from "./login-redirect";
-
-const APP_URL =
-  process.env.NEXT_PUBLIC_ADMIN_APP_URL || "https://admin.zoonk.com";
 
 async function LoginHandler() {
   const session = await getSession();
@@ -15,7 +12,7 @@ async function LoginHandler() {
     redirect("/");
   }
 
-  const callbackUrl = `${APP_URL}/auth/callback`;
+  const callbackUrl = `${getBaseUrl()}/auth/callback`;
   const authUrl = buildAuthLoginUrl({ callbackUrl });
 
   return <LoginRedirect url={authUrl} />;
