+++ apps/main/src/app/[locale]/(auth)/login/page.tsx
@@ -1,12 +1,10 @@
 import { getSession } from "@zoonk/core/users";
 import { FullPageLoading } from "@zoonk/ui/components/loading";
-import { buildAuthLoginUrl } from "@zoonk/utils/auth-url";
+import { buildAuthLoginUrl, getBaseUrl } from "@zoonk/utils/url";
 import { getLocale } from "next-intl/server";
 import { Suspense } from "react";
 import { redirect } from "@/i18n/navigation";
 import { LoginRedirect } from "./login-redirect";
-
-const APP_URL = process.env.NEXT_PUBLIC_MAIN_APP_URL || "https://zoonk.com";
 
 async function LoginHandler() {
   const session = await getSession();
@@ -16,7 +14,7 @@ async function LoginHandler() {
     redirect({ href: "/", locale });
   }
 
-  const callbackUrl = `${APP_URL}/${locale}/auth/callback`;
+  const callbackUrl = `${getBaseUrl()}/${locale}/auth/callback`;
   const authUrl = buildAuthLoginUrl({ callbackUrl, locale });
 
   return <LoginRedirect url={authUrl} />;
