+++ packages/utils/src/locale.ts
@@ -1,4 +1,5 @@
 export const SUPPORTED_LOCALES = ["en", "es", "pt"] as const;
+export const DEFAULT_LOCALE = "en";
 
 export type SupportedLocale = (typeof SUPPORTED_LOCALES)[number];
