+++ turbo.json
@@ -19,10 +19,8 @@
     "GOOGLE_CLIENT_ID",
     "GOOGLE_CLIENT_SECRET",
     "MAILER_API_KEY",
-    "NEXT_PUBLIC_ADMIN_APP_URL",
-    "NEXT_PUBLIC_AUTH_APP_URL",
-    "NEXT_PUBLIC_EDITOR_APP_URL",
-    "NEXT_PUBLIC_MAIN_APP_URL",
+    "VERCEL_ENV",
+    "VERCEL_URL",
     "OPENAI_API_KEY",
     "REVALIDATE_SECRET",
     "STRIPE_SECRET_KEY",
